Python/Trabalho-Final/requirements.txt
pygame
pillow
screeninfo